исправления по замечаниям ментора
