+++ packages/client/src/components/ServiceWorker/const.ts
@@ -2,13 +2,6 @@
 export const CACHE_NAME = `cache-name`;
 //список файлов для кеширования
 export const PAGES: string[] = [
-  "./logo.png",
-  "./pandocoverride.css",
-  "./style.css",
-  "./space.png",
-  "./reset.css",
-  "./index.css",
-  "./index.html",
-  "./game.html"
+  "./index.html"
 ];
